Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

The delay in granting su rights without the patch that this commit is about to revert is sometimes unbearable.
Therefore I would strongly prefer an actual fix of the issue before such plain revert as this.

Can you please test if this approach to "Don't wait" is fine regarding dumpstate? (I'm away on holidays, so I can't test it myself currently):

https://github.com/nadlabak/android_system_su/commit/bc4903c59672c4e26dc574415d53232d3246a63e

Patch-set: 1