fix storage format
The su binary in cm10 still uses the sqlite db to find if an app has
permission, but Superuser app has been updated and the protocol has changed.

So it fails to allow a certain app (like droidwall and openvpn) and it keeps
asking every time.

This only applies the patches of su-binary to reach the branch that has the new
protocol in it. (interestingly not origin/master, what a mess!)

Change-Id: I3eb6c19ad8521cc8b4aa69471b417412d7e35d51